requirements.txt
fastapi
uvicorn
chromadb
langchain
langchain-groq         
langchain-huggingface   
sentence-transformers   
langchain-community
python-dotenv